requirements.txt
django
selenium
djangorestframework
markdown
django-filter